+++ .gitignore
@@ -1,39 +1,16 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# testing
-/coverage
-
-# next.js
-/.next/
-/out/
-
-# production
-/build
-
-# misc
+# macOS oddities
 .DS_Store
-*.pem
 
-# debug
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-.pnpm-debug.log*
+# Generated files by hugo
+# /public/
+/resources/_gen/
+/assets/jsconfig.json
+hugo_stats.json
 
-# local env files
-.env*.local
+# Executable may be added to repository
+hugo.exe
+hugo.darwin
+hugo.linux
 
-# vercel
-.vercel
-
-# typescript
-*.tsbuildinfo
-next-env.d.ts
-
-# secrets
-secrets/
+# Temporary lock file while building
+/.hugo_build.lock
