+++ old/pages/404.tsx
