+++ old/styles/globals.css
@@ -1,5 +1,6 @@
 @import url("https://fonts.googleapis.com/css2?family=DM+Sans:ital,wght@0,500;0,700;1,400;1,500;1,700&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Urbanist:ital,wght@0,400;0,500;0,600;0,700;0,800;0,900;1,400&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Instrument+Sans:wght@400;500;600;700&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=League+Spartan:wght@100;200;300;400;500;600;700;800;900&display=swap");
 @import url("https://cdn.jsdelivr.net/npm/katex@0.16.3/dist/katex.min.css");
 
 @tailwind base;
@@ -7,7 +8,8 @@
 @tailwind utilities;
 
 body {
-  @apply bg-space;
+  /* @apply bg-black; */
+  background: #0c0d0e;
   @apply text-white;
 }
