+++ old/utils/default-seo.ts
